+++ packages/desktop/app/components/TwoTwoGrid.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const TwoTwoGrid = styled.div`
+  display: grid;
+
+  margin-top: 0.3rem;
+  grid-column-gap: 0.7rem;
+  grid-row-gap: 1.2rem;
+
+  grid-template-columns: 1fr 1fr;
+  align-items: start;
+`;
