+++ packages/desktop/app/forms/index.js
@@ -5,6 +5,6 @@ export { NewPatientForm } from './NewPatientForm';
 export { NewUserForm } from './NewUserForm';
 export { OngoingConditionForm } from './OngoingConditionForm';
 export { PatientIssueForm } from './PatientIssueForm';
-export { ImmunisationForm } from './ImmunisationForm';
+export { VaccineGivenForm } from './VaccineGivenForm';
 export { PatientCarePlanForm } from './PatientCarePlanForm';
 export { VitalsForm } from './VitalsForm';
